requirements.txt
Flask
python-dotenv 
googletrans 
pyttsx3 
mysql-connector-python
flask_session
googletrans==4.0.0-rc1
google-generativeai==0.8.3
werkzeug==3.0.4
whisper==1.1.10
openai-whisper==20231117
ffmpeg-python==0.2.0
moviepy
pydub